requirement.txt
pyaes==1.6.1
pyasn1==0.6.1
PyYAML==6.0.2
rsa==4.9.1
Telethon==1.40.0
